Add PersistentLineFactory and it's test class, reformat code
Created PersistentLineFactory which uses DB to create lines, created a
test class for it. Also created a new script testDataInsertion.sql to
use in the mentioned test class, updated the script for DB definition,
changed constructors in inMemory factories to make them resistent to
outside changes in parameters and reformated couple other classes for
SQLException handling.